requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
crypto==1.4.1
Django==5.2
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Naked==0.1.32
openapi-codec==1.3.2
psycopg2-binary==2.9.10
pycryptodome==3.22.0
PyJWT==2.9.0
PyYAML==6.0.2
requests==2.32.3
shellescape==3.8.1
simplejson==3.20.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
Pillow==10.4.0
drf-yasg[swagger]
pytest-django==4.5.2
pytest==7.4.0
pytest-cov==4.1.0
dj-database-url==2.3.0
django-cors-headers==4.7.0

asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==5.2
django-cors-headers==4.3.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-rest-swagger==2.2.0
drf-yasg[swagger]==1.21.7
idna==3.10
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Naked==0.1.32
openapi-codec==1.3.2
Pillow==10.4.0
psycopg2-binary==2.9.10
pycryptodome==3.22.0
PyJWT==2.9.0
PyYAML==6.0.2
requests==2.32.3
shellescape==3.8.1
simplejson==3.20.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
pytest==7.4.0
pytest-cov==4.1.0
pytest-django==4.5.2

